pmic: bd71892: Fix ADC measurement
The BD71892 has an ADC block, which can be used to measure voltages. The
measurement is started by writing to a "ADC KICK" -register. The
register is cleared when measurement is completed.

The BD71892 driver polled this register for measurement completion, via
I2C. Polling was done until I2C read returned a value which was smaller
than zero. The I2C read returns negative values only at error, so as
long as the I2C works, this polling will never be completed.

Correct logic is to end polling when read value is:
1. Zero, indicating measurement being done
Or
2. Negative value, indicating I2C error.

Fix the polling to be stopped when zero or negative value is read.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>